Solved lots of problems, put code together
* Fixed ElectionListener socket errors and made it an inner class
of Client, as to allow it to access their socket info.
* Client now waits for USER_RECONNECT_TIMEOUT seconds before deciding
that the server was closed for good ( to allow the replicas to perform
their election).
* When leader replica is elected, it opens a new connection with the
ElectionListener, which updates the client's server socket.
* new_server_address structure is no longer needed, since we can
obtaint he new server information from the opened connection.
* Fixed incorrect logic in Bully Algorithm where replicas could start
more than one election at a time.
* Not sure if election is 100% correct, there are some very rare bugs
where none of the replicas receieve each other's messages in time,
causing both to proclaim themselves leaders. This could be fixed by
increasing the election timeout, but that would imply the client has
to wait longer for the new main replica.
* Changed incorrect logic where sessions and their users were being
saved always as the main replica's socket. Solving this would require
re-engineering the entire program again, so I opted to ammend it.
* When a new leader is elected, it opens the new connection to all
the previous clients, and sends a "translation" - from old socket
to new socket - to every replica.
* Fixed missing "length" field in the message update packet, causing
messages to not be saved in other replicas' history file.
* Completed the election logic so the leaders now connect to the
previous clients upon being elected.
* Refactored a lot of code in ReplicaManager for readability.
* Almost there